update from pr comments
