+++ src/sentry/static/sentry/app/actionCreators/modal.tsx
@@ -1,8 +1,16 @@
 import React from 'react';
 import {css} from 'react-emotion';
+import {ModalHeader, ModalBody, ModalFooter} from 'react-bootstrap';
+``;
 
 import ModalActions, {ModalOptions} from 'app/actions/modalActions';
-import {RenderProps} from 'app/components/globalModal';
+
+export type RenderProps = {
+  closeModal: () => void;
+  Header: typeof ModalHeader;
+  Body: typeof ModalBody;
+  Footer: typeof ModalFooter;
+};
 
 /**
  * Show a modal
