+++ src/sentry/static/sentry/app/components/globalModal.tsx
@@ -4,19 +4,11 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import Reflux from 'reflux';
 import createReactClass from 'create-react-class';
-import {ModalHeader, ModalBody, ModalFooter} from 'react-bootstrap';
 
-import {closeModal} from 'app/actionCreators/modal';
+import {closeModal, RenderProps} from 'app/actionCreators/modal';
 import Confirm from 'app/components/confirm';
 import ModalStore from 'app/stores/modalStore';
 import {ModalOptions} from 'app/actions/modalActions';
-
-export type RenderProps = {
-  closeModal: () => void;
-  Header: typeof ModalHeader;
-  Body: typeof ModalBody;
-  Footer: typeof ModalFooter;
-};
 
 type Props = {
   children?: (renderProps: RenderProps) => React.ReactNode;
